+++ src/lib.rs
@@ -1,12 +1,12 @@
 use crate::error_enum::{AppResult, ParserErrorKind};
 use crate::parser::parse_all;
-use crate::validtor::validate_sturcture;
+use crate::validator::validate_sturcture;
 
 pub mod db_type;
 pub mod error_enum;
 pub mod tools;
 pub mod parser;
-pub mod validtor;
+pub mod validator;
 
 
 pub fn parse_file(input:&str) -> AppResult<()> {
