+++ src/validator/mod.rs
@@ -5,10 +5,10 @@ use crate::db_type::relation::Relation;
 use crate::db_type::table::{Column, Table};
 use crate::db_type::{GlobalDefinition, HashName};
 use crate::error_enum::{AppResult, ParserErrorKind};
-use crate::validtor::vailtor_schema::read_config;
-use crate::validtor::valid_enum::validate_enum;
-use crate::validtor::valid_relation::validate_relation;
-use crate::validtor::valid_table::validate_table;
+use crate::validator::vailtor_schema::read_config;
+use crate::validator::valid_enum::validate_enum;
+use crate::validator::valid_relation::validate_relation;
+use crate::validator::valid_table::validate_table;
 
 pub mod vailtor_schema;
 pub(crate) mod schema_enum;
