+++ src/validator/schema_table.rs
@@ -6,7 +6,7 @@ use serde::Deserialize;
 
 use crate::error_enum::schema_error::SchemaErrorKind;
 use crate::error_enum::{AppError, AppResult};
-use crate::validtor::vailtor_schema::compile_regex;
+use crate::validator::vailtor_schema::compile_regex;
 
 
 #[derive(Deserialize,Debug)]
