+++ src/validator/valid_enum.rs
@@ -1,8 +1,8 @@
 
 use crate::db_type::column_enum::ColumnEnum;
 use crate::error_enum::AppResult;
-use crate::validtor::schema_enum::{SchemaEnum, SchemaEnumRegex};
-use crate::validtor::vailtor_schema::validate_field_from_regex;
+use crate::validator::schema_enum::{SchemaEnum, SchemaEnumRegex};
+use crate::validator::vailtor_schema::validate_field_from_regex;
 
 
 fn validate_enum_item(
