+++ src/validator/valid_relation.rs
@@ -5,8 +5,8 @@ use crate::db_type::relation::Relation;
 use crate::db_type::table::Column;
 use crate::error_enum::schema_error::SchemaErrorKind;
 use crate::error_enum::AppResult;
-use crate::validtor::schema_relation::{SchemaRelation, SchemaRelationRegex};
-use crate::validtor::vailtor_schema::validate_field_from_regex;
+use crate::validator::schema_relation::{SchemaRelation, SchemaRelationRegex};
+use crate::validator::vailtor_schema::validate_field_from_regex;
 
 fn validate_relation_item(
     relation:&Relation,
